addContactHandler
