+++ server-api/db.json
@@ -19,6 +19,16 @@
       "id": "28fe91eb-3d69-4116-a582-f2493f561790",
       "name": "hêhhehe",
       "email": "giang6789@gmail.com"
+    },
+    {
+      "id": "7e3b1d9c-108f-454c-9480-668599e89f09",
+      "name": "giang ",
+      "email": "giang "
+    },
+    {
+      "id": "5bef7d8e-69a5-4e33-8567-a3b96a032b65",
+      "name": "hêhheh",
+      "email": "hrhrhrhr"
     }
   ]
 }
