+++ src/components/AddContact.js
@@ -1,50 +1,52 @@
-import React from "react";
+import React, {useState} from "react";
+import { useContactsCrud } from "../context/ContactsCrudContext";
+import { useNavigate } from "react-router-dom";
 
-class AddContact extends React.Component {
-  state = {
-    name: "",
-    email: "",
-  };
-  add = (e) => {
+const AddContact = () => {
+    const [name,setName] =useState ("");
+    const [email,setEmail] = useState("");
+    const {addContactHandler}  =useContactsCrud();
+    const navigate = useNavigate();
+
+ const add = (e) => {
     e.preventDefault();
-    if (this.state.name === "" || this.state.email === "") {
+    if (name === "" || email === "") {
       alert("All the fields are mandatory! ");
       return;
     }
-    this.props.addContactHandler(this.state);
-    this.setState({ name: "",  email:"" }); 
-    this.props.history.push("/");
+    addContactHandler({name,email});
+    setName("");
+    setEmail("");
+    navigate("/");
   };
-  render() {
     return (
       <div className="ui main">
         <h2> Add Contact</h2>
-        <form className="ui form" onSubmit={this.add}>
+        <form className="ui form" onSubmit={add}>
           <div className="field">
             <label>Name</label>
             <input
               type="text"
-              value={this.state.name}
+              value={name}
               name="name"
               placeholder="Name"
-              onChange={(e) => this.setState({ name: e.target.value })}
+              onChange={(e) => setName( e.target.value )}
             />
           </div>
           <div className=" field ">
             <label>Email</label>
             <input
               type="text"
-              value ={this.state.email}
+              value ={email}
               name="email"
               placeholder="Email"
-              onChange={(e) => this.setState({ email: e.target.value })}
+              onChange={(e) => setEmail( e.target.value )}
             />
           </div>
           <button className="ui button blue">Add</button>
         </form>
       </div>
     );
-  }
 }
 
 export default AddContact;
