+++ src/context/ContactsCrudContext.js
@@ -1,5 +1,6 @@
 import { createContext, useContext, useState } from "react";
 import api from "../api/contacts";
+import { v4 as uuid } from "uuid";
 
 const contactsCrudContext = createContext();
 
@@ -9,6 +10,18 @@ export function ContactsCrudContextProvider({children}) {
     const retrieveContacts = async () => {
       const response = await api.get("/contact");
       if(response.data)  setContacts(response.data);
+    };
+    // addContact
+    const addContactHandler = async (contact) => {
+      console.log("input contact,", contact);
+      const data = {
+        id: uuid(),
+        ...contact,
+      };
+      const response = await api.post("/contact", data);
+      console.log("output response", response);
+  
+      setContacts([response.data, ...contacts]);
     };
     // deleteContact
     const removeContactHandler = async (id) => {
@@ -22,7 +35,8 @@ export function ContactsCrudContextProvider({children}) {
   const value = {
     contacts,
     retrieveContacts,
-    removeContactHandler
+    removeContactHandler,
+    addContactHandler
   };
   return (
     <contactsCrudContext.Provider value={value}>
